#147: Sub-module inclusion for better browser compatibility
