+++ package.json
@@ -48,7 +48,7 @@
     "compile-doc": "tsc -p doc-gen",
     "compile": "yarn run compile-src && yarn run compile-test && npm run compile-doc",
     "lint": "tslint 'src/**/*.ts' --exclude 'src/**/*.d.ts' 'test/**/*.ts' --exclude 'test/**/*.d.ts' 'doc-gen/**/*.ts'",
-    "test": "cd test && mocha --require ts-node/register --full-trace .",
+    "test": "mocha --require ts-node/register --full-trace test",
     "build": "yarn run clean && yarn run compile && yarn run doc-gen",
     "start": "npm-run-all compile lint cover-test",
     "cover-test": "nyc yarn run test",
@@ -59,7 +59,7 @@
     "debug": "^4.0.1",
     "file-type": "9",
     "media-typer": "^0.3.0",
-    "strtok3": "^2.0.0",
+    "strtok3": "^2.1.0",
     "then-read-stream": "^1.3.0",
     "token-types": "^1.0.0"
   },
