+++ src/ogg/vorbis/Vorbis.ts
@@ -1,6 +1,7 @@
-import * as Token from "token-types";
-import {AttachedPictureType} from "../../id3v2/ID3v2";
-import {IPicture} from "../../index";
+import * as Token from 'token-types';
+
+import {AttachedPictureType} from '../../id3v2/ID3v2';
+import {IPicture} from '../../type';
 
 export interface IVorbisPicture extends IPicture {
   type: string
