+++ src/ogg/vorbis/VorbisTagMapper.ts
@@ -1,6 +1,6 @@
-import {INativeTagMap} from "../../common/GenericTagTypes";
-import {CommonTagMapper} from "../../common/GenericTagMapper";
-import {IRating, ITag} from "../../index";
+import {INativeTagMap} from '../../common/GenericTagTypes';
+import {CommonTagMapper} from '../../common/GenericTagMapper';
+import {IRating, ITag} from '../../type';
 
 /**
  * Vorbis tag mappings
