+++ src/riff/RiffInfoTagMap.ts
@@ -1,5 +1,5 @@
-import {INativeTagMap} from "../common/GenericTagTypes";
-import {CommonTagMapper} from "../common/GenericTagMapper";
+import {INativeTagMap} from '../common/GenericTagTypes';
+import {CommonTagMapper} from '../common/GenericTagMapper';
 
 /**
  * RIFF Info Tags; part of the EXIF 2.3
