+++ test/metadata-parsers.ts
@@ -1,7 +1,8 @@
 import * as fs from 'fs-extra';
 import * as mm from '../src';
+import {IAudioMetadata, IOptions} from '../src/type';
 
-type ParseFileMethod = (filePath: string, mimeType?: string, options?: mm.IOptions) => Promise<mm.IAudioMetadata>;
+type ParseFileMethod = (filePath: string, mimeType?: string, options?: IOptions) => Promise<IAudioMetadata>;
 
 interface IParser {
   description: string;
@@ -14,12 +15,12 @@ interface IParser {
 export const Parsers: IParser[] = [
   {
     description: 'parseFile',
-    initParser: (filePath: string, mimeType?: string, options?: mm.IOptions) => {
+    initParser: (filePath: string, mimeType?: string, options?: IOptions) => {
       return mm.parseFile(filePath, options);
     }
   }, {
     description: 'parseStream',
-    initParser: (filePath: string, mimeType?: string, options?: mm.IOptions) => {
+    initParser: (filePath: string, mimeType?: string, options?: IOptions) => {
       const stream = fs.createReadStream(filePath);
       return mm.parseStream(stream, mimeType, options).then(metadata => {
         stream.close();
@@ -28,7 +29,7 @@ export const Parsers: IParser[] = [
     }
   }, {
     description: 'parseBuffer',
-    initParser: (filePath: string, mimeType?: string, options?: mm.IOptions) => {
+    initParser: (filePath: string, mimeType?: string, options?: IOptions) => {
       const buffer = fs.readFileSync(filePath);
       return mm.parseBuffer(buffer, mimeType, options);
     }
