+++ test/test-file-ape.ts
@@ -2,6 +2,7 @@ import {assert} from 'chai';
 import * as mm from '../src';
 import * as path from 'path';
 import {Parsers} from './metadata-parsers';
+import {INativeTagDict} from '../src/type';
 
 const t = assert;
 
@@ -33,7 +34,7 @@ describe("Parse APE (Monkey's Audio)", () => {
     t.strictEqual(common.picture[1].data.length, 48658, 'common.picture 1 length');
   }
 
-  function checkNative(ape: mm.INativeTagDict) {
+  function checkNative(ape: INativeTagDict) {
     t.deepEqual(ape.ENSEMBLE, ['Audioslave']);
     t.deepEqual(ape.Artist, ['Audioslave', 'Chris Cornell']);
     t.strictEqual(ape['Cover Art (Front)'][0].data.length, 48658, 'raw cover art (front) length');
