+++ test/test-file-mp3.ts
@@ -1,6 +1,7 @@
 import {assert} from 'chai';
 import * as mm from '../src';
 import * as path from 'path';
+import {IFormat} from '../src/type';
 
 const t = assert;
 
@@ -27,7 +28,7 @@ describe("Parse MP3 files", () => {
 
     const filePath = path.join(__dirname, 'samples', "incomplete.mp3");
 
-    function checkFormat(format: mm.IFormat) {
+    function checkFormat(format: IFormat) {
       t.deepEqual(format.tagTypes, ['ID3v2.3', 'ID3v1'], 'format.tagTypes');
       t.approximately(format.duration, 61.73, 1 / 100, 'format.duration');
       t.strictEqual(format.dataformat, 'mp3', 'format.dataformat');
