+++ test/tsconfig.json
@@ -1,11 +1,7 @@
 {
+  "extends": "../tsconfig.json",
   "compilerOptions": {
-    "sourceMap": true,
-    "inlineSources": true,
-    "module": "commonjs",
-    "moduleResolution": "node",
-    "declaration": false,
-    "target": "es6"
+    "sourceMap": true
   }
 }
